[refactor] Create register module
- app.js에 있는 회원가입 로직을 모듈로 분리
