+++ db.json
@@ -123,6 +123,22 @@
           "complete": false
         }
       ]
+    },
+    {
+      "id": "sony1",
+      "info": {
+        "id": "sony1",
+        "pw": "123"
+      },
+      "todos": []
+    },
+    {
+      "id": "get",
+      "info": {
+        "id": "get",
+        "pw": "d"
+      },
+      "todos": []
     }
   ]
 }
